Created Product Detail Screen

+++ ApiMockData/src/main/java/com/ms/apimockdata/model/Reviews.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class ReviewsModel (
 
   @SerializedName("user_id" ) var userId  : Int?    = null,
-  @SerializedName("rating"  ) var rating  : Int?    = null,
+  @SerializedName("rating"  ) var rating  : Double?    = null,
   @SerializedName("comment" ) var comment : String? = null
 
 )
